--- src/contexts/AuthContext.tsx
@@ -1,45 +1,0 @@
-import { createContext, useEffect, useState, type ReactNode } from "react";
-
-interface User {
-  email: string;
-  name: string;
-}
-
-interface AuthContextType {
-  user: User | null;
-  setUser: (user: User) => void;
-  isLoggedIn: boolean;
-  logout: () => void;
-  isInitialized: boolean;
-}
-
-export const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const [user, setUserState] = useState<User | null>(null);
-  const [isInitialized, setIsInitialized] = useState(false);
-  
-  useEffect(() => {
-    const storedUser = localStorage.getItem("user");
-    if (storedUser) {
-        setUserState(JSON.parse(storedUser));
-    }
-    setIsInitialized(true);
-  }, []);
-
-  const setUser = (user: User) => {
-    setUserState(user);
-    localStorage.setItem("user", JSON.stringify(user));
-  };
-
-  const logout = () => {
-    setUserState(null);
-    localStorage.removeItem("user");
-  };
-
-  return (
-    <AuthContext.Provider value={{ user, setUser, isLoggedIn: !!user, logout, isInitialized }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
